Split apart config module
